+++ examples/manual/proguide/literals/intsuff.bas
@@ -6,7 +6,7 @@
 '' See Also: http://www.freebasic.net/wiki/wikka.php?wakka=ProPgLiterals
 '' --------
 
-Dim a As Integer = 123L
+Dim a As Long = 123L
 Dim b As UInteger = &h1234u
 Dim c As LongInt = 76543LL
 Dim d As ULongInt = &b1010101ULL
